+++ test/compute_test.py
@@ -158,6 +158,20 @@ class TestAzurermPy(unittest.TestCase):
         response = azurerm.get_compute_usage(self.access_token, self.subscription_id, self.location)
         self.assertTrue(len(response['value']) > 0)
 
+        # get vm instance view
+        print('Getting VM instance view')
+        response = azurerm.get_vm_instance_view(self.access_token, self.subscription_id, \
+            self.rgname, self.vmname)
+        # print(json.dumps(response, sort_keys=False, indent=2, separators=(',', ': ')))
+        self.assertEqual(response['statuses'][0]['displayStatus'], 'Creating')
+
+        # list vm instance views
+        print('Listing VM instance views')
+        response = azurerm.list_vm_instance_view(self.access_token, self.subscription_id, \
+            self.rgname)
+        # print(json.dumps(response, sort_keys=False, indent=2, separators=(',', ': ')))
+        self.assertTrue(len(response['value']) > 0)
+
         # delete VM
         print('Deleting VM: ' + self.vmname)
         response = azurerm.delete_vm(self.access_token, self.subscription_id, self.rgname, \